quicksend/quicksend/quicksend/requirements.txt
flask
werkzeug
Pillow
